:sparkles: feat: include zustand and refactor codes
