+++ package.json
@@ -18,7 +18,8 @@
     "react-hook-form": "^7.51.0",
     "react-router-dom": "^6.22.2",
     "uuid": "^9.0.1",
-    "yup": "^1.3.3"
+    "yup": "^1.3.3",
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
     "@types/node": "^20.11.22",
