--- src/interfaces/People.interfaces.ts
@@ -1,6 +1,0 @@
-interface IPeople {
-    data : string;
-    homens : number;
-    mulheres : number;
-    criancas : number;
-}
