+++ src/interfaces/people.interfaces.ts
@@ -1,0 +1,6 @@
+interface IPeople {
+    data : string;
+    homens : number;
+    mulheres : number;
+    criancas : number;
+}
